Updated license to MIT
The MIT License is a permissive open source license that allows users to use, modify, and redistribute your code for any purpose, as long as they keep the original copyright and license notice in the code. This license is a good choice if you want to give others as much freedom as possible to use your code.